backend/requirements.txt
# FastAPI
fastapi==0.128.0
uvicorn==0.40.0

# Pydantic
pydantic==2.12.5
pydantic-settings==2.12.0

# Supabase
supabase==2.27.1

# Utilities
python-dotenv==1.2.1
email-validator==2.3.0
python-jose==3.5.0
httpx==0.28.1

# Security
bcrypt==5.0.0
passlib==1.7.4
